[feat]: 학생증 인증 컴포넌트 분리
